Rename and format
