+++ app/src/main/java/app/pachli/components/notifications/NotificationsFragment.kt
@@ -144,7 +144,7 @@ class NotificationsFragment :
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
-        savedInstanceState: Bundle?
+        savedInstanceState: Bundle?,
     ): View {
         return inflater.inflate(R.layout.fragment_timeline_notifications, container, false)
     }
@@ -163,7 +163,7 @@ class NotificationsFragment :
 
         // Setup the SwipeRefreshLayout.
         binding.swipeRefreshLayout.setOnRefreshListener(this)
-        binding.swipeRefreshLayout.setColorSchemeColors(MaterialColors.getColor(binding.root, androidx.appcompat.R.attr.colorPrimary),)
+        binding.swipeRefreshLayout.setColorSchemeColors(MaterialColors.getColor(binding.root, androidx.appcompat.R.attr.colorPrimary))
 
         // Setup the RecyclerView.
         binding.recyclerView.setHasFixedSize(true)
